Added code to daemonize before initiating power cycle.
This feature is required so as to avoid a failure condition where the
item is shut off, and then we lose communications, stopping the program.
That sequence would leave the item shut off.  If the item was the router, we'd
be unable to communicate further to turn it back on.